requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
mysql-connector-python==8.2.0